+++ database/migrations/2026_08_06_205240_create_user_inventories_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->foreignId('ingredient_id')->constrained()->cascadeOnDelete();
-            $table->float('amount_left')->nullable();
+            $table->decimal('amount_left')->nullable();
             $table->enum('status', ['FULL', 'OPENED', 'LOW'])->nullable();
             $table->date('expiration_date')->nullable();
             $table->boolean('is_frozen')->default(false);
